Move functions in helpers.py into filters.py and make_polygons.py
